added multimeter projects for multiple backends
