+++ gradle.properties
@@ -41,6 +41,7 @@ slf4j_version=1.7.26
 
 # typesafe config couldn't be upgraded due to android restrictions
 typesafe_config_version=1.3.1
+micrometer_version=1.1.4
 
 # publishing
 spring_dependency_management_version=1.0.8.RELEASE
