+++ ktor-features/ktor-metrics-micrometer/build.gradle
@@ -1,15 +1,17 @@
 description = ''
 
-tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
-    kotlinOptions {
-        // micrometer uses jdk 1.8 features, so we need to use that version here
-        jvmTarget = "1.8"
+allprojects {
+    tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+        kotlinOptions {
+            // micrometer uses jdk 1.8 features, so we need to use that version here
+            jvmTarget = "1.8"
+        }
     }
 }
 
 kotlin {
     sourceSets.jvmMain.dependencies {
-        api group: 'io.micrometer', name: 'micrometer-core', version: '1.1.4'
+        api group: 'io.micrometer', name: 'micrometer-core', version: micrometer_version
         implementation project(':ktor-server:ktor-server-host-common')
     }
 }
