+++ ktor-features/ktor-metrics-micrometer/ktor-metrics-micrometer-dynatrace/build.gradle
@@ -1,0 +1,13 @@
+/*
+ * Copyright 2014-2019 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+description = ''
+
+kotlin {
+    sourceSets.jvmMain.dependencies {
+        api ("io.micrometer:micrometer-registry-dynatrace:$micrometer_version")
+        api project(':ktor-features:ktor-metrics-micrometer')
+    }
+}
+
