+++ settings.gradle
@@ -21,7 +21,7 @@ include ':ktor-server:ktor-server-tests'
 include ':ktor-server:ktor-server-host-common'
 include ':ktor-server:ktor-server-test-host'
 include ':ktor-server:ktor-server-jetty'
-include ':ktor-server:ktor-server-jetty:ktor-server-jetty-test-http2'
+//include ':ktor-server:ktor-server-jetty:ktor-server-jetty-test-http2'
 include ':ktor-server:ktor-server-servlet'
 include ':ktor-server:ktor-server-tomcat'
 include ':ktor-server:ktor-server-netty'
@@ -62,6 +62,22 @@ include ':ktor-features:ktor-gson'
 include ':ktor-features:ktor-jackson'
 include ':ktor-features:ktor-metrics'
 include ':ktor-features:ktor-metrics-micrometer'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-appoptics'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-atlas'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-datadog'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-dynatrace'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-elastic'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-ganglia'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-graphite'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-humio'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-influx'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-jmx'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-kairosdb'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-newrelic'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-prometheus'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-signalfx'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-statsd'
+include ':ktor-features:ktor-metrics-micrometer:ktor-metrics-micrometer-wavefront'
 include ':ktor-features:ktor-server-sessions'
 include ':ktor-features:ktor-locations'
 include ':ktor-features:ktor-websockets'
